# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "NTI API",
        "description": "REST API pre platformu Národná technologická iniciatíva (NTI) — registrácie, prihlášky, hodnotenie, projekty, dokumenty, CMS a administráciu.",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "/api",
            "description": "API server"
        }
    ],
    "paths": {
...
        "/applications": {
            "get": {
                "tags": [
                    "Applications"
                ],
                "summary": "List applications (paginated)",
                "description": "Возвращает постраничный список заявок",
                "operationId": "53a774224a5bb97dd3d5e1c1d0055cd9",
                "responses": {
                    "200": {
                        "description": "Paginated list of applications"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Applications"
                ],
                "summary": "Create a draft application for the current team and an open call",
                "description": "Создаёт черновик заявки для текущей команды",
                "operationId": "c8746709b6a737b932ae65fd891652ad",
                "responses": {
                    "201": {
                        "description": "Application created (draft)"
                    },
                    "403": {
                        "description": "Not authorized to create applications"
                    },
                    "422": {
                        "description": "Validation error"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/applications/{application}": {
            "get": {
                "tags": [
                    "Applications"
                ],
                "summary": "Get a single application with team, organization, answers and pairing submissions",
                "description": "Возвращает детали одной заявки со связанными данными",
                "operationId": "31160acc92874271b270a5e8211410f7",
                "parameters": [
                    {
                        "name": "application",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Application detail"
                    },
                    "403": {
                        "description": "Not authorized to view this application"
                    },
                    "404": {
                        "description": "Application not found"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Applications"
                ],
                "summary": "Update an application's answers, pairing submissions or organization",
                "description": "Обновляет ответы, документы для парного отбора или организацию заявки",
                "operationId": "abea4f6d9a3d4ec0ec92743ab38aad49",
                "parameters": [
                    {
                        "name": "application",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Application updated"
                    },
                    "403": {
                        "description": "Not authorized to update this application"
                    },
                    "422": {
                        "description": "Validation error"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Applications"
                ],
                "summary": "Delete an application",
                "description": "Удаляет заявку",
                "operationId": "96ae1ef06f193366689c077802986df3",
                "parameters": [
                    {
                        "name": "application",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Application deleted"
                    },
                    "403": {
                        "description": "Not authorized to delete this application"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/applications/{application}/submit": {
            "post": {
                "tags": [
                    "Applications"
                ],
                "summary": "Submit a draft application for committee review",
                "description": "Подаёт черновик заявки на рассмотрение комиссией",
                "operationId": "010fc1a431178d8567a60d2911a2ebbf",
                "parameters": [
                    {
                        "name": "application",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Application submitted"
                    },
                    "403": {
                        "description": "Not authorized to submit this application"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/applications/{application}/decide": {
            "post": {
                "tags": [
                    "Applications"
                ],
                "summary": "[Committee] Approve, reject or request supplements for a submitted application",
                "description": "Фиксирует решение комиссии по поданной заявке",
                "operationId": "d425853b83e31a3d71b1f9cffd591149",
                "parameters": [
                    {
                        "name": "application",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Decision recorded, application updated"
                    },
                    "403": {
                        "description": "Not authorized to decide on this application"
                    },
                    "422": {
                        "description": "Validation error"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "sanctum": {
                "type": "http",
                "scheme": "bearer",
                "bearerFormat": "Token",
                "description": "Enter your Sanctum token (without \"Bearer \" prefix)"
            }
        }
    },
...
    "security": [
        {
            "sanctum": []
        }
    ]
}